% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RGTvcftools.R
\name{callBafChanges}
\alias{callBafChanges}
\title{Call sites showing BAF changes}
\usage{
callBafChanges(baftable, cref, csample, foldchange = 1.2, maxref = 0.05,
  minAF = 0.15)
}
\arguments{
\item{baftable}{data.frame with variants and allelic frequencies}

\item{cref}{identifier for column with reference BAF (e.g. normal tissue)}

\item{csample}{identifier for column with matched sample BAF (e.g. tumor tissue)}

\item{foldchange}{minimum fold.change threshold to call mutation}

\item{maxref}{numeric, maximum allowed BAF in the reference sample}

\item{minAF}{numeric, minimum BAF required in the sample}
}
\description{
This function performs very crude mutation calling based on BAFs. The calling is
depends only on fold change and minimum AF thresholds.
}

