% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RGTvcftools.R
\name{callTrioDeNovo}
\alias{callTrioDeNovo}
\title{Call sites de novo mutations from family trio data}
\usage{
callTrioDeNovo(baftable, cref1 = "mother", cref2 = "father",
  csample = "child", thesaurus = FALSE, foldchange = 1.2, maxref = 0.05,
  minsample = 0.15, mincov = 10, exceptcovchr1 = c(),
  exceptcovchr2 = c())
}
\arguments{
\item{baftable}{data.frame with variants and allelic frequency}

\item{cref1,}{string, code refererring to one of the matched controls, i.e. one of the parents.
The function assumes the baftable contains columns such as
cref1.thesaurus.BAF, etc as output by GeneticThesaurus software}

\item{cref2}{string, similar to cref1}

\item{csample}{string, code referring to the child sample. The function assumes the baftable
contains columns such as csample.thesaurus.BAF, etc. as output
by GeneticThesaurus software.}

\item{thesaurus}{boolean, set TRUE to use thesaurus adjusted BAF and coverage estimates}

\item{foldchange}{numeric, required fold change in allelic frequency}

\item{maxref}{numeric, maximal allelic frequency allowed in the reference samples}

\item{minsample}{numeric, minimum allelic frequency required in the child sample}

\item{mincov}{numeric, minimum coverage required in the reference samples}

\item{exceptcovchr1}{vector of chromosome names that should be excluded from the
minimum coverage requirement (e.g. set chrY for a female sample)}

\item{exceptcovchr2}{vector, similar to exceptcovchr1}
}
\description{
This function performs crude mutation calling based on BAF. It searches for variants
in a child sample that are not present in either parent. The calculation depends
on changes in AF between child and parents.
}

