% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RGTvcftools.R
\name{compareVariants}
\alias{compareVariants}
\title{Compare a set of variants (hits) with a set of expected variants (known sites).}
\usage{
compareVariants(known, hits, links = NULL)
}
\arguments{
\item{known}{data frame of known/expected variants; should contain column names (chr, position)}

\item{hits}{data frame of hits/observed variants, should contain column names (chr, position)}

\item{links}{four-column table with thesaurus links as obtained by function getLinksFromFile().
These links are assumed to correspond to variants given via the data frame hits.}
}
\description{
This function compares two sets of variants and reports the number of true positives,
false positives and false negatives.
}

