% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RGTvcftools.R
\name{isIndel}
\alias{isIndel}
\title{Determines whether a variant is a substitution or an indel}
\usage{
isIndel(rb, ab)
}
\arguments{
\item{rb}{vector of reference bases}

\item{ab}{vector of alternative bases}
}
\description{
Given ref and alt alleles, this function returns a logical vector describing
whether or not the variants represent indels. The function regards entries like
rb="X" ab="Y" as not an indel. The function regards rb="X" and ab="Y,Z" as not an indel.
Other types of ref/alt combinations are reported as indels.
}

