% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RGTvcftools.R
\name{print.RGeneticThesaurusSet}
\alias{print.RGeneticThesaurusSet}
\title{Display information about an RGeneticThesaurus object}
\usage{
\method{print}{RGeneticThesaurusSet}(x, ...)
}
\arguments{
\item{x}{an object of class RGeneticThesaurus to print}

\item{...}{not used}
}
\description{
Display information about an RGeneticThesaurus object
}

