% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RGTvcftools.R
\name{readClustersFromFile}
\alias{readClustersFromFile}
\title{Read a table with thesaurus clusters from file}
\usage{
readClustersFromFile(ff)
}
\arguments{
\item{ff}{filename}
}
\description{
This function is just a wrapper for read.table(). It is included in the package for
completeness so that all thesaurus-related files can be accessed with similar function names.
}

