% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RGTvcftools.R
\name{readLinksFromFile}
\alias{readLinksFromFile}
\title{Read thesaurus links from a file}
\usage{
readLinksFromFile(ff, n = 65536)
}
\arguments{
\item{ff}{filename of vtf}

\item{n}{number of lines. The function will read the vtf in multiple chunks, each chunk
containing at most these many lines.}
}
\description{
This function reads a vtf file from disk and returns a data.frame. The data.frame is
organized into six columns. Each row represents one thesaurus link and has an origin and
destination.
}

