% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RGTvcftools.R
\name{sortvcf}
\alias{sortvcf}
\title{Sort a data frame of variants by chromosome and position}
\usage{
sortvcf(vcf, chrorder = NULL)
}
\arguments{
\item{vcf}{data frame with variants; should contain columns (chr, position)}

\item{chrorder}{vector of chromosome names; order of this vector will determine sort order}
}
\description{
Sort a data frame by chromosome and position. (Note, this function could be faster
using data.table. But this implementation avoids dependencies)
}

