% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RGTvcftools.R
\name{readThesaurusSet}
\alias{readThesaurusSet}
\title{Read a set of variants annotated using GeneticThesaurus}
\usage{
readThesaurusSet(variantsfile, n = 4096, ignorelines = c("thesaurushard",
  "thesaurusmany"), getcolumns = 10, withindels = FALSE)
}
\arguments{
\item{variantsfile}{filename for a vcf file annotated with the genetic thesaurus.
This function assumes variantsfile ends with vcf (vcf.gz), and then looks for files
with similar name with extensions vtf (or vtf.gz) and baf.tsv (or baf.tsv.gz)}

\item{n}{number of lines to read (used by readVariantsFromFile and readLinksFromFile}

\item{ignorelines}{filter codes to remove from variant list}

\item{getcolumns}{integer; this is passed on to function readVariantsFromFile}

\item{withindels}{logical; determines whether variant list should include insertions/deletions}
}
\description{
Read a set of variants annotated using GeneticThesaurus
}

